+++ noodles-fasta/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.1 - 2021-12-09
+
+### Fixed
+
+  * fasta: Sync dependencies.
+
 ## 0.5.0 - 2021-12-02
 
 ### Added